requirements.txt
requests>=2.0.0
selenium>=3.0.0
pandas
pillow